requirements.txt
numpy
python-binance
TA-Lib
websocket_client
pytz
